message delivery method change

+++ app/src/main/java/com/marslan/chatarneca/data/SharedRepository.kt
@@ -1,41 +1,28 @@
 package com.marslan.chatarneca.data
 
 import androidx.lifecycle.LiveData
-import com.marslan.chatarneca.data.chatdb.EntityChat
-import com.marslan.chatarneca.data.messagedb.EntityMessage
 
 class SharedRepository(private val sharedDao: SharedDao) {
 
     fun getChatMessage(id : Int) : LiveData<List<EntityMessage>> = sharedDao.getChatMessage(id)
 
-    suspend fun newMessage(entityMessage: EntityMessage, toID: String){
-        sharedDao.newMessage(entityMessage)
-        newChat(entityMessage,toID)
-        sharedDao.updateChatLastMessage(!entityMessage.isRead,entityMessage.text,entityMessage.date,entityMessage.chatID)
-    }
+    val allChatWithLastMessage: LiveData<List<EntityMessage>> = sharedDao.allChatWithLastMessage()
+
+    suspend fun newMessage(entityMessage: EntityMessage){sharedDao.newMessage(entityMessage)}
+
+    fun deleteMessage(entityMessage: EntityMessage){sharedDao.deleteMessage(entityMessage)}
+
+    suspend fun updateMessage(entityMessage: EntityMessage){sharedDao.updateMessage(entityMessage)}
 
     val readAllChat: LiveData<List<EntityChat>> = sharedDao.getChat()
 
-    private suspend fun newChat(entityMessage: EntityMessage,toID: String){
-        sharedDao.newChat(
-            EntityChat(
-                entityMessage.chatID,
-                "chat",
-                toID,
-                !entityMessage.isRead,
-                entityMessage.text,
-                entityMessage.date
-            )
-        )
-    }
-    fun updateChat(read: Boolean,id: Int){
-        sharedDao.updateChatRead(read,id)
-    }
+    suspend fun newChat(entityChat: EntityChat){sharedDao.newChat(entityChat)}
 
     fun deleteChat(entityChat: EntityChat){
-        sharedDao.deleteChat(entityChat)
-        sharedDao.deleteMessage(entityChat.chatID)
+        if(entityChat.users.split("%").size <= 2)
+            sharedDao.deleteChat(entityChat)
+        sharedDao.deleteMessageWithChat(entityChat.id)
     }
 
-    suspend fun updateChatName(entityChat: EntityChat){sharedDao.updateChat(entityChat)}
+    suspend fun updateChat(entityChat: EntityChat){sharedDao.updateChat(entityChat)}
 }
